修正64位cgns网格转换时的问题，将第三方库及相应头文件从版本库中移除，用户编译时需要自行编译第三方库。消除源代码中的部分warning

+++ codes/database/include/DataBaseIO.h
@@ -170,7 +170,7 @@ void HXWrite( TIO * tio, HXVector< T > & field )
 template < typename T >
 void HXWrite( fstream * file, HXVector< T > & field )
 {
-    Int nElement = field.size();
+    Int nElement = static_cast<int> (field.size());
     HXWrite( file, & field[ 0 ], nElement );
 }
